Add FP16 support to LocalParamWithInput.

+++ neurite/tf/layers.py
@@ -1513,7 +1513,7 @@ class LocalParamWithInput(Layer):
     def call(self, x):
         # want the x variable for it's keras properties and the batch.
         xslice = K.batch_flatten(x)[:, 0:1]
-        b = xslice * tf.zeros((1,)) + tf.ones((1,))
+        b = xslice * tf.zeros((1,), dtype=x.dtype) + tf.ones((1,), dtype=x.dtype)
         # b = K.batch_flatten(0 * x)[:, 0:1] + 1
         params = K.flatten(self.kernel * self.biasmult)[tf.newaxis, ...]
         return K.reshape(K.dot(b, params), [-1, *self.shape])
